Moved sub-classes of "NotifyIcon", "ToolStripDropDownItem", "ToolStripSplitButton" and "ToolStripStatusLabel" classes to their own files
