+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolStripSplitButton.cs
@@ -9,7 +9,6 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Windows.Forms.Design;
 using System.Windows.Forms.Layout;
-using static Interop;
 
 namespace System.Windows.Forms
 {
@@ -599,217 +598,6 @@ namespace System.Windows.Forms
         internal virtual bool ShouldSerializeDropDownButtonWidth()
         {
             return (DropDownButtonWidth != DefaultDropDownButtonWidth);
-        }
-
-        /// <summary>
-        ///  This class represents the item to the left of the dropdown [ A |v]  (e.g the "A")
-        ///  It exists so that we can use our existing methods for text and image layout
-        ///  and have a place to stick certain state information like pushed and selected
-        ///  Note since this is NOT an actual item hosted on the ToolStrip - it wont get things
-        ///  like MouseOver, wont be laid out by the ToolStrip, etc etc.  This is purely internal
-        ///  convenience.
-        /// </summary>
-        private class ToolStripSplitButtonButton : ToolStripButton
-        {
-            private readonly ToolStripSplitButton _owner;
-
-            public ToolStripSplitButtonButton(ToolStripSplitButton owner)
-            {
-                this._owner = owner;
-            }
-
-            public override bool Enabled
-            {
-                get
-                {
-                    return _owner.Enabled;
-                }
-                set
-                {
-                    // do nothing
-                }
-            }
-
-            public override ToolStripItemDisplayStyle DisplayStyle
-            {
-                get
-                {
-                    return _owner.DisplayStyle;
-                }
-                set
-                {
-                    // do nothing
-                }
-            }
-
-            public override Padding Padding
-            {
-                get
-                {
-                    return _owner.Padding;
-                }
-                set
-                {
-                    // do nothing
-                }
-            }
-
-            public override ToolStripTextDirection TextDirection
-            {
-                get
-                {
-                    return _owner.TextDirection;
-                }
-            }
-
-            public override Image Image
-            {
-                get
-                {
-                    if ((_owner.DisplayStyle & ToolStripItemDisplayStyle.Image) == ToolStripItemDisplayStyle.Image)
-                    {
-                        return _owner.Image;
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-                set
-                {
-                    // do nothing
-                }
-            }
-
-            public override bool Selected
-            {
-                get
-                {
-                    if (_owner != null)
-                    {
-                        return _owner.Selected;
-                    }
-                    return base.Selected;
-                }
-            }
-
-            public override string Text
-            {
-                get
-                {
-                    if ((_owner.DisplayStyle & ToolStripItemDisplayStyle.Text) == ToolStripItemDisplayStyle.Text)
-                    {
-                        return _owner.Text;
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-                set
-                {
-                    // do nothing
-                }
-            }
-        }
-
-        /// <summary>
-        ///  This class performs internal layout for the "split button button" portion of a split button.
-        ///  Its main job is to make sure the inner button has the same parent as the split button, so
-        ///  that layout can be performed using the correct graphics context.
-        /// </summary>
-        private class ToolStripSplitButtonButtonLayout : ToolStripItemInternalLayout
-        {
-            readonly ToolStripSplitButton _owner;
-
-            public ToolStripSplitButtonButtonLayout(ToolStripSplitButton owner) : base(owner.SplitButtonButton)
-            {
-                this._owner = owner;
-            }
-
-            protected override ToolStripItem Owner
-            {
-                get { return _owner; }
-            }
-
-            protected override ToolStrip ParentInternal
-            {
-                get
-                {
-                    return _owner.ParentInternal;
-                }
-            }
-            public override Rectangle ImageRectangle
-            {
-                get
-                {
-                    Rectangle imageRect = base.ImageRectangle;
-                    // translate to ToolStripItem coordinates
-                    imageRect.Offset(_owner.SplitButtonButton.Bounds.Location);
-                    return imageRect;
-                }
-            }
-
-            public override Rectangle TextRectangle
-            {
-                get
-                {
-                    Rectangle textRect = base.TextRectangle;
-                    // translate to ToolStripItem coordinates
-                    textRect.Offset(_owner.SplitButtonButton.Bounds.Location);
-                    return textRect;
-                }
-            }
-        }
-
-        internal class ToolStripSplitButtonUiaProvider : ToolStripDropDownItemAccessibleObject
-        {
-            private readonly ToolStripSplitButtonExAccessibleObject _accessibleObject;
-
-            public ToolStripSplitButtonUiaProvider(ToolStripSplitButton owner) : base(owner)
-            {
-                _accessibleObject = new ToolStripSplitButtonExAccessibleObject(owner);
-            }
-
-            public override void DoDefaultAction()
-            {
-                _accessibleObject.DoDefaultAction();
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                return _accessibleObject.GetPropertyValue(propertyID);
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                return _accessibleObject.IsPatternSupported(patternId);
-            }
-
-            internal override void Expand()
-            {
-                DoDefaultAction();
-            }
-
-            internal override void Collapse()
-            {
-                _accessibleObject.Collapse();
-            }
-
-            internal override UiaCore.ExpandCollapseState ExpandCollapseState
-            {
-                get
-                {
-                    return _accessibleObject.ExpandCollapseState;
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                return _accessibleObject.FragmentNavigate(direction);
-            }
         }
     }
 }
